create hook for race querystring
